Modified demo console log event to be more verbose. Changed some calendar property names to be 'showingMonth' instead of 'currentMonth' to separate selection functionality from view functionality. Changed 'setMonth' naming to 'selectMonth' naming for my concise conveyance of function. Modified sl-calendar-month to yield and removed the idea that it should know its own name since that should come from calendar and locale. Commented out several properties in sl-calendar-month. Modified the sl-calendar-day click action to send the date as well as any content or data it may have.
Separated changeMonth style functions from setMonth style functions so
that one will use the other. Added a selectDate action for when days are
clicked. Sl-calendar now has a separation of selectedDate vs viewingDate.
showingMonth and showingYear are based on viewingDate. Active months and
years are now based on the selectedDate and not arbitrary data points.

Added sl-calendar as subcomponent on sl-date-picker. Added an action to
sl-date-picker to listen for the calendar date selection. Added locale
awareness to sl-date-picker. Commented old datepicker plugin
initialization code in sl-date-picker.